requirements.txt
torch==2.2.0
tqdm>=4.57.0
boto3>=1.17.15
jsonschema>=3.2.0
Deprecated>=1.2.11
scipy>=1.6.1
matplotlib>=3.3.4
psutil>=5.8.0
tensorboard>=2.4.1
setuptools>=21.0.0
coverage~=5.3.1
torchvision>=0.10.0
sphinx~=4.0.2
sphinx-rtd-theme
torchmetrics==0.8
hydra-core>=1.2.0
omegaconf
onnxruntime==1.15.1
onnx==1.15.0
pillow>=5.3.0,!=8.3
pip-tools>=6.12.1
pyparsing==2.4.5
einops==0.3.2
pycocotools==2.0.6
protobuf==3.20.3
treelib==1.6.1
termcolor==1.1.0
packaging>=20.4
# not directly required, pinned by Snyk to avoid a vulnerability
wheel>=0.38.0
# not directly required, pinned by Snyk to avoid a vulnerability
pygments>=2.7.4
stringcase>=1.2.0
numpy<=1.24.2
rapidfuzz
json-tricks==3.16.1
onnxsim>=0.4.3,<1.0
data-gradients~=0.3.1
albumentations~=1.3
modal==0.62.192
